+++ src/cc_search_chats/semantic/__init__.py
@@ -6,6 +6,8 @@ from cc_search_chats.semantic.model import (
     chunk_passages,
     embed_passages,
     embed_query,
+    local_model_revision,
+    model_output_scope,
 )
 
 __all__ = [
@@ -14,4 +16,6 @@ __all__ = [
     "chunk_passages",
     "embed_passages",
     "embed_query",
+    "local_model_revision",
+    "model_output_scope",
 ]